ajout dependances Makefile + norme parser + corrections bugs
