+++ srcs/main.c
@@ -27,6 +27,35 @@ void	init_lexer(t_lexer *lexer)
 	lexer->curr_flag = NULL;
 }
 
+int		lexer_routine(char **line, t_lexer *lexer)
+{
+	set_signal_ign();
+	init_lexer(lexer);
+	if (!ft_lexer(line, lexer))
+	{
+		ft_strdel(line);
+		return (0);
+	}
+	if (cfg_shell()->debug)
+		ft_lstiter(lexer->token_lst, print_debug);
+	ft_strdel(line);
+	return (1);
+}
+
+int		parser_routine(t_lexer *lexer,t_parser *parser)
+{
+	if (!ft_parser(lexer, parser)
+	|| (parser->state == S_PARSER_TABLE_START
+	&& !parser->table))
+	{
+		ft_lstdel(&parser->table, del_cmd_table);
+		ft_lstdel(&lexer->token_lst, del_token);
+		return (0);
+	}
+	ft_lstdel(&lexer->token_lst, del_token);
+	return (1);
+}
+
 int		main(int ac, char **av, char **env)
 {
 	ssize_t		ret;
@@ -44,15 +73,9 @@ int		main(int ac, char **av, char **env)
 			continue ;
 		else if (line && (!line[0] || ft_strcmp("exit\n", line) == 0))
 			break ;
-		set_signal_ign();
-		init_lexer(&lexer);
-		if (!ft_lexer(&line, &lexer))
+		if (!lexer_routine(&line, &lexer)
+		|| !parser_routine(&lexer, &parser))
 			continue ;
-		if (debug)
-			ft_lstiter(lexer.token_lst, print_debug);
-		if (!ft_parser(&lexer, &parser))
-			continue ;
-		ft_lstdel(&lexer.token_lst, del_token);
 		if (parser.state != S_PARSER_SYNTAX_ERROR)
 		{
 			if (debug)
