+++ srcs/parser/parser.c
@@ -44,216 +44,6 @@ int		p_add_redir_delim(t_token* token, t_parser *parser)
 	return (1);
 }
 
-void		p_init_start_state(int (*table_builder[10][17])(t_token *, t_parser *))
-{
-	table_builder[S_PARSER_TABLE_START][TOKEN] = syn_err;
-	table_builder[S_PARSER_TABLE_START][WORD] = p_cmd_name;
-	table_builder[S_PARSER_TABLE_START][ASSIGNMENT_WORD] = p_add_assign;
-	table_builder[S_PARSER_TABLE_START][NEWLINE] = syn_err;
-	table_builder[S_PARSER_TABLE_START][IO_NUMBER] = p_add_io_num;
-	table_builder[S_PARSER_TABLE_START][AND_IF] = syn_err;
-	table_builder[S_PARSER_TABLE_START][AMP] = syn_err;
-	table_builder[S_PARSER_TABLE_START][OR_IF] = syn_err;
-	table_builder[S_PARSER_TABLE_START][PIPE] = syn_err;
-	table_builder[S_PARSER_TABLE_START][SEMI] = syn_err;
-	table_builder[S_PARSER_TABLE_START][LESS] = p_add_redir;
-	table_builder[S_PARSER_TABLE_START][DLESS] = p_add_redir;
-	table_builder[S_PARSER_TABLE_START][GREAT] = p_add_redir;
-	table_builder[S_PARSER_TABLE_START][DGREAT] = p_add_redir;
-	table_builder[S_PARSER_TABLE_START][LESSAND] = p_add_redir;
-	table_builder[S_PARSER_TABLE_START][GREATAND] = p_add_redir;
-	table_builder[S_PARSER_TABLE_START][DLESSDASH] = p_add_redir;
-}
-
-void		p_init_cmd_wait_state(int (*table_builder[10][17])(t_token *, t_parser *))
-{
-	table_builder[S_PARSER_CMD_START][TOKEN] = syn_err;
-	table_builder[S_PARSER_CMD_START][WORD] = p_cmd_name;
-	table_builder[S_PARSER_CMD_START][ASSIGNMENT_WORD] = p_add_assign;
-	table_builder[S_PARSER_CMD_START][NEWLINE] = p_skip;
-	table_builder[S_PARSER_CMD_START][IO_NUMBER] = p_add_io_num;
-	table_builder[S_PARSER_CMD_START][AND_IF] = syn_err;
-	table_builder[S_PARSER_CMD_START][AMP] = syn_err;
-	table_builder[S_PARSER_CMD_START][OR_IF] = syn_err;
-	table_builder[S_PARSER_CMD_START][PIPE] = p_add_cmd;
-	table_builder[S_PARSER_CMD_START][SEMI] = p_set_start_state;
-	table_builder[S_PARSER_CMD_START][LESS] = p_add_redir;
-	table_builder[S_PARSER_CMD_START][DLESS] = p_add_redir;
-	table_builder[S_PARSER_CMD_START][GREAT] = p_add_redir;
-	table_builder[S_PARSER_CMD_START][DGREAT] = p_add_redir;
-	table_builder[S_PARSER_CMD_START][LESSAND] = p_add_redir;
-	table_builder[S_PARSER_CMD_START][GREATAND] = p_add_redir;
-	table_builder[S_PARSER_CMD_START][DLESSDASH] = p_add_redir;
-}
-
-void		p_init_args_wait_state(int (*table_builder[10][17])(t_token *, t_parser *))
-{
-	table_builder[S_PARSER_CMD_ARGS][TOKEN] = syn_err;
-	table_builder[S_PARSER_CMD_ARGS][WORD] = p_add_arg;
-	table_builder[S_PARSER_CMD_ARGS][ASSIGNMENT_WORD] = p_add_assign_arg;
-	table_builder[S_PARSER_CMD_ARGS][NEWLINE] = p_set_start_state;
-	table_builder[S_PARSER_CMD_ARGS][IO_NUMBER] = p_add_io_num;
-	table_builder[S_PARSER_CMD_ARGS][AND_IF] = p_add_and_or;
-	table_builder[S_PARSER_CMD_ARGS][AMP] = p_add_amp;
-	table_builder[S_PARSER_CMD_ARGS][OR_IF] = p_add_and_or;
-	table_builder[S_PARSER_CMD_ARGS][PIPE] = p_add_cmd;
-	table_builder[S_PARSER_CMD_ARGS][SEMI] = p_set_start_state;
-	table_builder[S_PARSER_CMD_ARGS][LESS] = p_add_redir;
-	table_builder[S_PARSER_CMD_ARGS][DLESS] = p_add_redir;
-	table_builder[S_PARSER_CMD_ARGS][GREAT] = p_add_redir;
-	table_builder[S_PARSER_CMD_ARGS][DGREAT] = p_add_redir;
-	table_builder[S_PARSER_CMD_ARGS][LESSAND] = p_add_redir;
-	table_builder[S_PARSER_CMD_ARGS][GREATAND] = p_add_redir;
-	table_builder[S_PARSER_CMD_ARGS][DLESSDASH] = p_add_redir;
-}
-
-void		p_init_redir_state(int (*table_builder[10][17])(t_token *, t_parser *))
-{
-	table_builder[S_PARSER_REDIR][TOKEN] = syn_err;
-	table_builder[S_PARSER_REDIR][WORD] = p_file_name;
-	table_builder[S_PARSER_REDIR][ASSIGNMENT_WORD] = syn_err;
-	table_builder[S_PARSER_REDIR][NEWLINE] = syn_err;
-	table_builder[S_PARSER_REDIR][IO_NUMBER] = syn_err;
-	table_builder[S_PARSER_REDIR][AND_IF] = syn_err;
-	table_builder[S_PARSER_REDIR][AMP] = syn_err;
-	table_builder[S_PARSER_REDIR][OR_IF] = syn_err;
-	table_builder[S_PARSER_REDIR][PIPE] = syn_err;
-	table_builder[S_PARSER_REDIR][SEMI] = syn_err;
-	table_builder[S_PARSER_REDIR][LESS] = syn_err;
-	table_builder[S_PARSER_REDIR][DLESS] = syn_err;
-	table_builder[S_PARSER_REDIR][GREAT] = syn_err;
-	table_builder[S_PARSER_REDIR][DGREAT] = syn_err;
-	table_builder[S_PARSER_REDIR][LESSAND] = syn_err;
-	table_builder[S_PARSER_REDIR][GREATAND] = syn_err;
-	table_builder[S_PARSER_REDIR][DLESSDASH] = syn_err;
-}
-
-void		p_init_assign_state(int (*table_builder[10][17])(t_token *, t_parser *))
-{
-	table_builder[S_PARSER_ASSIGN][TOKEN] = syn_err;
-	table_builder[S_PARSER_ASSIGN][WORD] = p_assign_val;
-	table_builder[S_PARSER_ASSIGN][ASSIGNMENT_WORD] = syn_err;
-	table_builder[S_PARSER_ASSIGN][NEWLINE] = p_set_start_state;
-	table_builder[S_PARSER_ASSIGN][IO_NUMBER] = syn_err;
-	table_builder[S_PARSER_ASSIGN][AND_IF] = p_add_and_or;
-	table_builder[S_PARSER_ASSIGN][AMP] = p_add_amp;
-	table_builder[S_PARSER_ASSIGN][OR_IF] = p_add_and_or;
-	table_builder[S_PARSER_ASSIGN][PIPE] = p_add_cmd;
-	table_builder[S_PARSER_ASSIGN][SEMI] = p_set_start_state;
-	table_builder[S_PARSER_ASSIGN][LESS] = syn_err;
-	table_builder[S_PARSER_ASSIGN][DLESS] = syn_err;
-	table_builder[S_PARSER_ASSIGN][GREAT] = syn_err;
-	table_builder[S_PARSER_ASSIGN][DGREAT] = syn_err;
-	table_builder[S_PARSER_ASSIGN][LESSAND] = syn_err;
-	table_builder[S_PARSER_ASSIGN][GREATAND] = syn_err;
-	table_builder[S_PARSER_ASSIGN][DLESSDASH] = syn_err;
-}
-
-void		p_init_io_nbr_state(int (*table_builder[10][17])(t_token *, t_parser *))
-{
-	table_builder[S_PARSER_IO_NUMBER][TOKEN] = syn_err;
-	table_builder[S_PARSER_IO_NUMBER][WORD] = syn_err;
-	table_builder[S_PARSER_IO_NUMBER][ASSIGNMENT_WORD] = syn_err;
-	table_builder[S_PARSER_IO_NUMBER][NEWLINE] = syn_err;
-	table_builder[S_PARSER_IO_NUMBER][IO_NUMBER] = syn_err;
-	table_builder[S_PARSER_IO_NUMBER][AND_IF] = syn_err;
-	table_builder[S_PARSER_IO_NUMBER][AMP] = syn_err;
-	table_builder[S_PARSER_IO_NUMBER][OR_IF] = syn_err;
-	table_builder[S_PARSER_IO_NUMBER][PIPE] = syn_err;
-	table_builder[S_PARSER_IO_NUMBER][SEMI] = syn_err;
-	table_builder[S_PARSER_IO_NUMBER][LESS] = p_add_redir;
-	table_builder[S_PARSER_IO_NUMBER][DLESS] = p_add_redir;
-	table_builder[S_PARSER_IO_NUMBER][GREAT] = p_add_redir;
-	table_builder[S_PARSER_IO_NUMBER][DGREAT] = p_add_redir;
-	table_builder[S_PARSER_IO_NUMBER][LESSAND] = p_add_redir;
-	table_builder[S_PARSER_IO_NUMBER][GREATAND] = p_add_redir;
-	table_builder[S_PARSER_IO_NUMBER][DLESSDASH] = p_add_redir;
-}
-
-void		p_init_delim_state(int (*table_builder[10][17])(t_token *, t_parser *))
-{
-	table_builder[S_PARSER_DELIM][TOKEN] = syn_err;
-	table_builder[S_PARSER_DELIM][WORD] = p_add_redir_delim;
-	table_builder[S_PARSER_DELIM][ASSIGNMENT_WORD] = syn_err;
-	table_builder[S_PARSER_DELIM][NEWLINE] = syn_err;
-	table_builder[S_PARSER_DELIM][IO_NUMBER] = syn_err;
-	table_builder[S_PARSER_DELIM][AND_IF] = syn_err;
-	table_builder[S_PARSER_DELIM][AMP] = syn_err;
-	table_builder[S_PARSER_DELIM][OR_IF] = syn_err;
-	table_builder[S_PARSER_DELIM][PIPE] = syn_err;
-	table_builder[S_PARSER_DELIM][SEMI] = syn_err;
-	table_builder[S_PARSER_DELIM][LESS] = syn_err;
-	table_builder[S_PARSER_DELIM][DLESS] = syn_err;
-	table_builder[S_PARSER_DELIM][GREAT] = syn_err;
-	table_builder[S_PARSER_DELIM][DGREAT] = syn_err;
-	table_builder[S_PARSER_DELIM][LESSAND] = syn_err;
-	table_builder[S_PARSER_DELIM][GREATAND] = syn_err;
-	table_builder[S_PARSER_DELIM][DLESSDASH] = syn_err;
-}
-
-void		p_init_arg_assign_state(int (*table_builder[10][17])(t_token *, t_parser *))
-{
-	table_builder[S_PARSER_ARG_ASSIGN][TOKEN] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][WORD] = p_assign_join;
-	table_builder[S_PARSER_ARG_ASSIGN][ASSIGNMENT_WORD] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][NEWLINE] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][IO_NUMBER] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][AND_IF] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][AMP] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][OR_IF] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][PIPE] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][SEMI] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][LESS] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][DLESS] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][GREAT] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][DGREAT] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][LESSAND] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][GREATAND] = syn_err;
-	table_builder[S_PARSER_ARG_ASSIGN][DLESSDASH] = syn_err;
-}
-
-void		p_init_andif_pipe_state(int (*table_builder[10][17])(t_token *, t_parser *))
-{
-	table_builder[S_PARSER_ANDIF_PIPE][TOKEN] = syn_err;
-	table_builder[S_PARSER_ANDIF_PIPE][WORD] = p_cmd_name;
-	table_builder[S_PARSER_ANDIF_PIPE][ASSIGNMENT_WORD] = p_add_assign;
-	table_builder[S_PARSER_ANDIF_PIPE][NEWLINE] = p_skip;
-	table_builder[S_PARSER_ANDIF_PIPE][IO_NUMBER] = p_add_io_num;
-	table_builder[S_PARSER_ANDIF_PIPE][AND_IF] = syn_err;
-	table_builder[S_PARSER_ANDIF_PIPE][AMP] = syn_err;
-	table_builder[S_PARSER_ANDIF_PIPE][OR_IF] = syn_err;
-	table_builder[S_PARSER_ANDIF_PIPE][PIPE] = syn_err;
-	table_builder[S_PARSER_ANDIF_PIPE][SEMI] = syn_err;
-	table_builder[S_PARSER_ANDIF_PIPE][LESS] = p_add_redir;
-	table_builder[S_PARSER_ANDIF_PIPE][DLESS] = p_add_redir;
-	table_builder[S_PARSER_ANDIF_PIPE][GREAT] = p_add_redir;
-	table_builder[S_PARSER_ANDIF_PIPE][DGREAT] = p_add_redir;
-	table_builder[S_PARSER_ANDIF_PIPE][LESSAND] = p_add_redir;
-	table_builder[S_PARSER_ANDIF_PIPE][GREATAND] = p_add_redir;
-	table_builder[S_PARSER_ANDIF_PIPE][DLESSDASH] = p_add_redir;
-}
-
-void		p_init_syn_err_state(int (*table_builder[10][17])(t_token *, t_parser *))
-{
-	table_builder[S_PARSER_SYNTAX_ERROR][TOKEN] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][WORD] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][ASSIGNMENT_WORD] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][NEWLINE] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][IO_NUMBER] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][AND_IF] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][AMP] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][OR_IF] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][PIPE] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][SEMI] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][LESS] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][DLESS] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][GREAT] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][DGREAT] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][LESSAND] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][GREATAND] = p_skip;
-	table_builder[S_PARSER_SYNTAX_ERROR][DLESSDASH] = p_skip;
-}
-
 void		init_parser(t_parser *parser)
 {
 	parser->state = S_PARSER_TABLE_START;
